Move ThreatModel to threat package
